Rotas, botoes e testes para mudar status de Prato

+++ app/controllers/dishes_controller.rb
@@ -38,6 +38,7 @@ class DishesController < UserController
     found_dish = Dish.find(dish_id)
 
     verify_dish_owner(found_dish)
+    return if performed?
 
     if @dish.update(get_dish_params)
       redirect_to restaurant_dish_path(@restaurant, @dish), notice: 'Prato alterado com sucesso'
@@ -52,11 +53,44 @@ class DishesController < UserController
     found_dish = Dish.find(dish_id)
 
     verify_dish_owner(found_dish)
+    return if performed?
 
     if @dish.destroy
       redirect_to restaurant_dishes_path, notice: "Prato deletado com sucesso!"
     else
       flash.now[:alert] = "Não foi possível deletar o Prato"
+      render 'show', status: :unprocessable_entity
+    end
+  end
+
+  def deactivate
+    dish_id = params[:id]
+    found_dish = Dish.find(dish_id)
+    
+    verify_dish_owner(found_dish)
+    return if performed?
+
+    if @dish.active? && @dish.inactive!
+      flash[:notice] = 'Sucesso! Prato desativado'
+      redirect_to restaurant_dish_path(@restaurant, @dish)
+    else
+      flash.now[:notice] = 'Ocorreu um erro ao desativar o Prato'
+      render 'show', status: :unprocessable_entity
+    end
+  end
+
+  def activate
+    dish_id = params[:id]
+    found_dish = Dish.find(dish_id)
+
+    verify_dish_owner(found_dish)
+    return if performed?
+
+    if @dish.inactive? && @dish.active!
+      flash[:notice] = 'Sucesso! Prato ativado'
+      redirect_to restaurant_dish_path(@restaurant, @dish)
+    else
+      flash.now[:notice] = 'Ocorreu um erro ao ativar o Prato'
       render 'show', status: :unprocessable_entity
     end
   end
@@ -72,13 +106,13 @@ class DishesController < UserController
 
   def verify_dish_owner(found_dish)
     if found_dish.nil? 
-      return redirect_to root_path, alert: 'Ops! Prato não encontrado'
+      redirect_to root_path, alert: 'Ops! Prato não encontrado'
     end
 
     if found_dish.restaurant.id != @restaurant.id
       redirect_to root_path, alert: 'Ops! Você não tem acesso a pratos que não são do seu restaurante'
-    else
-        @dish = found_dish
     end
+
+    @dish = found_dish
   end
 end
